+++ index.js
@@ -5,5 +5,3 @@
 import './shim.js'
 import './native.js'
 import Config from 'react-native-config'
-
-console.disableYellowBox = !!Config.TEST_REACT_NATIVE;
