+++ src/lib/hooks/useConnectionGun.e2e.js
@@ -1,0 +1,1 @@
+export default () => true
