+++ src/lib/hooks/useConnectionWeb3.e2e.js
@@ -1,0 +1,1 @@
+export default () => true
